new file:   app/controllers/admin/categories_controller.rb 	deleted:    app/controllers/categories_controller.rb 	modified:   app/models/category.rb 	renamed:    app/views/categories/new.html.erb -> app/views/admin/categories/_form.html.erb 	new file:   app/views/admin/categories/edit.html.erb 	new file:   app/views/admin/categories/index.html.erb 	new file:   app/views/admin/categories/new.html.erb 	renamed:    app/views/categories/show.html.erb -> app/views/admin/categories/show.html.erb 	deleted:    app/views/categories/edit.html.erb 	deleted:    app/views/categories/index.html.erb 	modified:   config/routes.rb
