--- app/controllers/categories_controller.rb
@@ -1,47 +1,0 @@
-class CategoriesController < ApplicationController
-  def index
-    @categories = Category.all
-  end
-
-  def show
-    @category = Category.find(params[:id])
-  end
-
-  def new
-    @category = Category.new
-  end
-
-  def create
-    @category = Category.new(category_params)
-
-    if @category.save
-      redirect_to categories_path, :notice => "categoria creada"
-    else
-      render "new"
-    end
-  end
-
-  def edit
-    @category = Category.find(params[:id])
-  end
-
-  def update
-    @category = Category.find(params[:id])
-
-    if @category.update_attributes(category_params)
-      redirect_to categories_path, :notice => "Categoria actualizada"
-      else
-        render "edit"
-    end
-  end
-
-  def destroy
-    @category = Category.find(params[:id])
-    @category.destroy
-    redirect_to categories_path, :notice => "Categoria borrada"
-  end
-
-  def category_params
-      params.require(:category).permit(:nombre_categoria)
-  end
-end
