+++ app/models/category.rb
@@ -1,2 +1,3 @@
 class Category < ApplicationRecord
+  validates :nombre_categoria, presence: true, uniqueness: true
 end
