+++ app/views/admin/categories/edit.html.erb
@@ -1,0 +1,2 @@
+<h1>Editar categoria</h1>
+<%= render "form" %>
