+++ app/views/admin/categories/index.html.erb
@@ -1,0 +1,12 @@
+<h1>Categorias</h1>
+
+<% @categories.each do |category| %>
+   <h2><%= link_to category.nombre_categoria, admin_category_path(category) %></h2>
+   <p>
+   <%= link_to "Editar", edit_admin_category_path(category) %>
+   <%= link_to "Eliminar", admin_category_path(category), :confirm => "Confirme que desea eliminar la categoria", :method => :delete %>
+   </p>
+   <hr />
+<% end %>
+
+<p><%= link_to "Nueva categoria", new_admin_category_path %></p>
