+++ app/views/admin/categories/show.html.erb
@@ -1,3 +1,4 @@
 
 <p><%= @category.nombre_categoria %></p>
 <small><%= @category.created_at %></small>
+<p><%= link_to "regresar", admin_categories_path %></p>
