--- app/views/categories/new.html.erb
@@ -1,11 +1,0 @@
-<h1>Crear una nueva categoria</h1>
-
-<%= form_for @category do |f| %>
-  <p>
-    <%= f.label :nombre_categoria %><br />
-    <%= f.text_field :nombre_categoria %>
-  </p>
-  <p>
-    <%= f.submit "crear" %>
-  </p>
-<% end %>
