+++ config/routes.rb
@@ -1,14 +1,9 @@
 Rails.application.routes.draw do
+  namespace :admin do
 
   resources :categories
 
-  get 'categories/index'
-
-  get 'categories/new'
-
-  get 'categories/edit'
-
-  get 'categories/show'
+  end
 
   # For details on the DSL available within this file, see http://guides.rubyonrails.org/routing.html
 end
